Changed structure of testing folder to simplify orientation for adding testing of new components. Fixed intends in various files.
