+++ tests/UI/Component/ComponentHelperTest.php
