+++ tests/UI/Component/Counter/CounterFactoryTest.php
