+++ tests/UI/Component/Counter/CounterTest.php
@@ -3,7 +3,7 @@
 /* Copyright (c) 2016 Richard Klees <richard.klees@concepts-and-training.de> Extended GPL, see docs/LICENSE */
 
 require_once("libs/composer/vendor/autoload.php");
-require_once(__DIR__."/Base.php");
+require_once(__DIR__."/../../Base.php");
 
 use \ILIAS\UI\Component as C;
