+++ tests/UI/Component/Glyph/GlyphFactoryTest.php
