+++ tests/UI/Crawler/Fixture/ComponentsTreeFixture/Component1/Component12/Factory.php
@@ -9,14 +9,14 @@ use ILIAS\UI\Component as C;
  * Some Random Comment
  */
 interface ProperEntry {
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component1\component12\component121
-     */
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component1\component12\component121
+	 */
 
-    public function component121();
+	public function component121();
 }
