+++ tests/UI/Crawler/Fixture/Fixture.php
@@ -3,36 +3,36 @@
 
 class ProperEntryFixture
 {
-    public $properEntryYamlArray = array(
-        array(
-            "description" => array(
-                "purpose" => "Description of Purpose",
-                "composition" => "Description of Composition with line break",
-                "effect" => "Effect Description on one line",
-                "rivals" => array(
-                    "icon" => "Icon Description"
-                )
+	public $properEntryYamlArray = array(
+		array(
+			"description" => array(
+				"purpose" => "Description of Purpose",
+				"composition" => "Description of Composition with line break",
+				"effect" => "Effect Description on one line",
+				"rivals" => array(
+					"icon" => "Icon Description"
+				)
 
-            ),
-            "background"=>"\"Some wild background with quotes\" and links <a href='http:test'>test</a>",
-            "featurewiki"=> array("http://www.ilias.de/docu/goto_docu_wiki_wpage_3208_1357.html"),
-            "rules"=>array(
-                "usage"=>array(
-                    1 => "Usage Rule 1",
-                    2 => "Usage Rule 2",
-                    3 => "Usage Rule 3 multi line"
-                ),
-                "style"=>array(
-                    4 => "Style Rule"
-                )
-            ),
-            "namespace"=>"\\ILIAS\\UI\\Crawler\\Fixture\\ProperEntry",
-            "function_name"=>"properEntry"
+			),
+			"background"=>"\"Some wild background with quotes\" and links <a href='http:test'>test</a>",
+			"featurewiki"=> array("http://www.ilias.de/docu/goto_docu_wiki_wpage_3208_1357.html"),
+			"rules"=>array(
+				"usage"=>array(
+					1 => "Usage Rule 1",
+					2 => "Usage Rule 2",
+					3 => "Usage Rule 3 multi line"
+				),
+				"style"=>array(
+					4 => "Style Rule"
+				)
+			),
+			"namespace"=>"\\ILIAS\\UI\\Crawler\\Fixture\\ProperEntry",
+			"function_name"=>"properEntry"
 
-        )
-    );
+		)
+	);
 
-    public $properEntryYamlString =
+	public $properEntryYamlString =
 "description:
   purpose: >
       Description of Purpose
