+++ tests/UI/Crawler/Fixture/IdenticalEntriesFactory.php
@@ -9,23 +9,23 @@ use ILIAS\UI\Component as C;
  * Some Random Comment
  */
 interface IdenticalEntries {
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  tests\UI\Crawler\Fixture\Component2\component
-     */
-    public function component();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  tests\UI\Crawler\Fixture\Component2\component
+	 */
+	public function component();
 
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  tests\UI\Crawler\Fixture\Component2\component
-     */
-    public function component();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  tests\UI\Crawler\Fixture\Component2\component
+	 */
+	public function component();
 }
