+++ tests/UI/Crawler/Fixture/IdenticalNamesFactory.php
@@ -9,23 +9,23 @@ use ILIAS\UI\Component as C;
  * Some Random Comment
  */
 interface IdenticalNamesEntry {
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  \ILIAS\UI\Crawler\Fixture\ProperEntry
-     */
-    public function component();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  \ILIAS\UI\Crawler\Fixture\ProperEntry
+	 */
+	public function component();
 
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component2\Factory
-     */
-    public function component();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component2\Factory
+	 */
+	public function component();
 }
