+++ tests/UI/Crawler/Fixture/NoClosingDescriptionFactory.php
@@ -9,22 +9,22 @@ use ILIAS\UI\Component as C;
  * Some Random Comment
  */
 interface ProperEntry {
-    /**
-     * ---
-     * description:
-     * rules:
-     *
-     * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component1\Factory
-     */
-    public function component1();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 *
+	 * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component1\Factory
+	 */
+	public function component1();
 
-    /**
-     * ---
-     * description:
-     * rules:
-     * ---
-     *
-     * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component2\Factory
-     */
-    public function component2();
+	/**
+	 * ---
+	 * description:
+	 * rules:
+	 * ---
+	 *
+	 * @return  tests\UI\Crawler\Fixture\ComponentsTreeFixture\Component2\Factory
+	 */
+	public function component2();
 }
