+++ tests/UI/Crawler/Fixture/NoReturnValueEntry.php
@@ -9,37 +9,37 @@ use ILIAS\UI\Component as C;
  * Some Random Comment
  */
 interface noReturnValueEntry {
-    /**
-     * ---
-     *
-     * description:
-     *   purpose: >
-     *       Description of Purpose
-     *   composition: >
-     *       Description of Composition
-     *       with line break
-     *   effect: Effect Description on one line
-     *   rival:
-     *       icon: >
-     *           Icon Description
-     *
-     * background: >
-     *     "Some Wid Background with quotes"
-     *
-     *     and links <a href='http:test'>test</a>
-     * featurewiki:
-     *       - http://www.ilias.de/docu/goto_docu_wiki_wpage_3208_1357.html
-     *
-     * rules:
-     *   usage:
-     *       1: Usage Rule 1
-     *       2: Usage Rule 2
-     *       3: >
-     *         Usage Rule 3 multi line
-     *   style:
-     *       4: Style Rule
-     * ---
-     *
-     */
-    public function noReturnValueEntry();
+	/**
+	 * ---
+	 *
+	 * description:
+	 *   purpose: >
+	 *       Description of Purpose
+	 *   composition: >
+	 *       Description of Composition
+	 *       with line break
+	 *   effect: Effect Description on one line
+	 *   rival:
+	 *       icon: >
+	 *           Icon Description
+	 *
+	 * background: >
+	 *     "Some Wid Background with quotes"
+	 *
+	 *     and links <a href='http:test'>test</a>
+	 * featurewiki:
+	 *       - http://www.ilias.de/docu/goto_docu_wiki_wpage_3208_1357.html
+	 *
+	 * rules:
+	 *   usage:
+	 *       1: Usage Rule 1
+	 *       2: Usage Rule 2
+	 *       3: >
+	 *         Usage Rule 3 multi line
+	 *   style:
+	 *       4: Style Rule
+	 * ---
+	 *
+	 */
+	public function noReturnValueEntry();
 }
