+++ tests/UI/MainFactoryTest.php
@@ -2,11 +2,14 @@
 
 require_once 'tests/UI/AbstractFactoryTest.php';
 
+/**
+ * Class MainFactoryTest
+ */
 class MainFactoryTest extends AbstractFactoryTest {
 	public $kitchensink_info_settings = array(
-			"card"		=> array("context" => false)
-			, "deck"			=> array("context" => false)
-    );
+		"card"		=> array("context" => false)
+		, "deck"			=> array("context" => false)
+	);
 
 	public $factory_title = 'ILIAS\\UI\\Factory';
 
@@ -18,6 +21,9 @@ class MainFactoryTest extends AbstractFactoryTest {
 		// Nothing to test here.
 	}
 
+	/**
+	 * @return string
+	 */
 	protected function get_regex_factory_namespace() {
 		return "\\\\ILIAS\\\\UI\\\\Component";
 	}
