+++ tests/UI/Renderer/ilIndependentTemplate.php
